Mouse Click Changes
Check Readme file for more information at the top where it talks about
Click controls

also just a fyi, this was a bitch to do.

New Class: EmberChamberFlags. Has some pretty sweet code, might port to
cpp later (more dynamic version)

Wtf is with vel pinch, not suppose to be like that. it's not releasing
correctly