Blocking (anim only, no actual block)
Holding rightclick goes to block
Letting go of rightclick cancels block
Tapping rightclick will cause it to blend in and then straight out, so
no change

You can 'block' during attacks, but it'll cancel both attack and block
(attack cancels into block, and then the attack timer ends which causes
an animation reset making block go to idle)

Changes to block can be done under AttackFramework. Cause i'm using a
structure for this, block setup is in a function and not in
defaultproperties. Block setup is under the function "SetUpBlockPacket",
near the top of the functions in that page. Feel free to rearrange as
need be to make it more comfortable