Multiplayer Part IV
I'm going to learn all my roman numerals

Figured out how to replicate to all clients! Though im not sure if this
is the correct way, but it works!!!

However, while the animation 'name' gets replicated, for some reason the
mods don't get replicated. so animation is done w/ 0s time frame (if a
constant time frame is used in ClientPlayAnim instead, everything
replicates np)

TODO: Copy nether and optimize. It'll be tedious but server reliable.
Essentially convert all 'names' to integers, and never copy over mods.
Everyone already has mods,so why waste bandwidth? Idea is to send a
simple int to clients. When they receive it, they find the anim
associated and mods with it and play it on target pawn. Boom win.