UnSetup fix from custom cooking
If you cook the game through one of the bat files, the first time you
run it it'll work fine, but the second+ times you run it, UnSetup will
crash making UDK to crash. Replace UnSetup.exe in the Binaries folder
with this one to fix

DO NOT replace the UnSetup in your UDK.exe directory. ONLY from the
cooked game. You'll know when to use it.