Multiplayer Part XIII
Changed way damage works

There are 15 tracers per sword like before, but they are now split up in
3 groups (hilt, middle, tip) consisting of 5 tracers

if a tracer from a group is hit, the entire group is 'activated' and
will do group damage

so if tip = 1 damage, middle = 2, and hilt = 3, if the tip and middle
group were activated, a total of 3 damage would happen (1+2)

This is only for multiplayer. Local still uses damage per tracer.

You can see and change damage values under ReplicateDamage function in
EmberPawn