[Big Update] New Sword Management
Sword.uc is now obsolete. It is replaced by two files

alot of work is still needed. but for now it's working really good and
much better than previous. with lots more improvement to come

EmberSword_Model is the visual sword. Also contains the physics assets.
It is the thing that goes swing swing

EmberSword_Calculations is where everything is controlled. It'll do
tracer calculations on sockets, or on manual vectors (for lightsaber for
instance)

So now instead of having three duplicates of swords that each do
calculations independently, there is now a single central entity that
does the calculations and thinking for ALL swords (theoretically, can do
it for an infinite amount of swords). All the EmberSword_Model does is
give a visual representation back to the player

in Attack Framework, replaced all heavy shit with mediums. so stances 2
and 3 are both mediums (stance 3 uses lightsaber though)

Tracers are drawn slightly smoother now, even more so inbetween attacks.
It should feel VERY fluid now.

We are no longer using TracerEnd notifier in animations. That was
causing issues because of the blending. Instead I hacked together a way
to setup the tracers based on the non-windup animation. Works suitably.

Stuff's rather complicated, I don't know if I'll make a visual graph for
it (rather, I don't have the time)

Known issue: Chamber doesn't work well with new system so it's been
disabled.